Clean up boot log, auto install pakz archives (#260)

+++ workspace/tg5040/install/boot.sh
@@ -4,6 +4,7 @@
 PLATFORM="tg5040"
 SDCARD_PATH="/mnt/SDCARD"
 UPDATE_PATH="$SDCARD_PATH/MinUI.zip"
+PAKZ_PATH="$SDCARD_PATH/*.pakz"
 SYSTEM_PATH="$SDCARD_PATH/.system"
 
 echo userspace > /sys/devices/system/cpu/cpu0/cpufreq/scaling_governor
@@ -11,28 +12,46 @@ CPU_PATH=/sys/devices/system/cpu/cpu0/cpufreq/scaling_setspeed
 CPU_SPEED_PERF=2000000
 echo $CPU_SPEED_PERF > $CPU_PATH
 
-##REmove Old Led Daemon
-LCDAEMON_PATH="/etc/LedControl"
+##Remove Old Led Daemon
+if [ -f "/etc/LedControl" ]; then
+	rm -Rf "/etc/LedControl"
+fi
+if [ -f "/etc/init.d/lcservice" ]; then
+	/etc/init.d/lcservice disable
+	rm /etc/init.d/lcservice
+fi
 
-cd $(dirname "$0")
-rm -R $LCDAEMON_PATH
-/etc/init.d/lcservice disable
-rm /etc/init.d/lcservice
+export LD_LIBRARY_PATH=/usr/trimui/lib:$LD_LIBRARY_PATH
+export PATH=/usr/trimui/bin:$PATH
+
+TRIMUI_MODEL=`strings /usr/trimui/bin/MainUI | grep ^Trimui`
+if [ "$TRIMUI_MODEL" = "Trimui Brick" ]; then
+	DEVICE="brick"
+fi
+
+# leds_off
+echo 0 > /sys/class/led_anim/max_scale
+
+# generic NextUI package install
+for pakz in $PAKZ_PATH
+do
+	echo $pakz
+	cd $(dirname "$0")/$PLATFORM
+	./show.elf ./$DEVICE/installing.png
+
+	./unzip -o -d "$SDCARD_PATH" "$pakz" # >> $pakz.txt
+	rm -f "$pakz"
+
+	# run postinstall if present
+	if [ -f $SDCARD_PATH/post_install.sh ]; then
+		$SDCARD_PATH/post_install.sh # > $pakz_post.txt
+		rm -f $SDCARD_PATH/post_install.sh
+	fi
+done
 
 # install/update
 if [ -f "$UPDATE_PATH" ]; then 
 	echo ok
-	export LD_LIBRARY_PATH=/usr/trimui/lib:$LD_LIBRARY_PATH
-	export PATH=/usr/trimui/bin:$PATH
-
-	# leds_off
-	echo 0 > /sys/class/led_anim/max_scale
-	
-	TRIMUI_MODEL=`strings /usr/trimui/bin/MainUI | grep ^Trimui`
-	if [ "$TRIMUI_MODEL" = "Trimui Brick" ]; then
-		DEVICE="brick"
-	fi
-	
 	cd $(dirname "$0")/$PLATFORM
 	if [ -d "$SYSTEM_PATH" ]; then
 		./show.elf ./$DEVICE/updating.png
@@ -52,7 +71,6 @@ if [ -f "$UPDATE_PATH" ]; then
 	if [ -f $SYSTEM_PATH/$PLATFORM/bin/install.sh ]; then
 		$SYSTEM_PATH/$PLATFORM/bin/install.sh # &> $SDCARD_PATH/log.txt
 	fi
-	
 fi
 
 LAUNCH_PATH="$SYSTEM_PATH/$PLATFORM/paks/MinUI.pak/launch.sh"
